cleaned up step_1, moved it, renamed it, and moved static references to sql_info
